Moved Matched Map logic to it's own Fragment class. Added Info Window for the Other User. Changed the relative layout for the messaging bar in activity_contact to linearlayout
