+++ app/src/main/res/layout/activity_contact.xml
@@ -1,76 +1,71 @@
-    <RelativeLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context="com.lunchdash.lunchdash.activities.ContactActivity">
+
+    <LinearLayout
+        android:id="@+id/info"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <ImageView
+            android:id="@+id/ivUserImage"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:background="@color/com_facebook_blue"
+            tools:image="@drawable/ic_launcher" />
+
+        <TextView
+            android:id="@+id/tvContactText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:gravity="center_horizontal"
+            android:textIsSelectable="true"
+            android:textSize="12sp"
+            android:textColor="@color/primary_material_dark"
+            tools:text="'Name' is ready for an awesome lunch!\nGet in touch with him/her!" />
+    </LinearLayout>
+
+
+    <FrameLayout
+        android:id="@+id/flContactMap"
+        android:layout_below="@+id/info"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        tools:context="com.lunchdash.lunchdash.activities.ContactActivity">
+        android:layout_height="200dp" />
 
-        <LinearLayout
-            android:id="@+id/info"
-            android:layout_width="wrap_content"
+    <ListView
+        android:id="@+id/lvChat"
+        android:layout_below="@+id/flContactMap"
+        android:layout_above="@+id/llSend"
+        android:layout_width="wrap_content"
+        android:transcriptMode="alwaysScroll"
+        android:layout_height="match_parent" />
+
+
+    <LinearLayout
+        android:id="@+id/llSend"
+        android:layout_alignParentBottom="true"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <EditText
+            android:id="@+id/etMessage"
+            android:layout_width="0dp"
+            android:layout_weight="0.7"
             android:layout_height="wrap_content"
-            android:orientation="horizontal">
-            <ImageView
-                android:id="@+id/ivUserImage"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:background="@color/com_facebook_blue"
-                tools:image="@drawable/ic_launcher" />
+            android:hint="@string/message_hint" />
 
-            <TextView
-                android:id="@+id/tvContactText"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:textIsSelectable="true"
-                android:textSize="12sp"
-                android:textColor="@color/primary_material_dark"
-                tools:text="'Name' is ready for an awesome lunch!\nGet in touch with him/her!" />
-        </LinearLayout>
-
-
-        <fragment
-            android:id="@+id/mvContactMap"
-            android:layout_below="@+id/info"
-            class="com.google.android.gms.maps.SupportMapFragment"
-            android:layout_width="match_parent"
-            android:layout_height="200dp"/>
-
-        <ListView
-            android:id="@+id/lvChat"
-            android:layout_below="@+id/mvContactMap"
-            android:layout_above="@+id/llSend"
-            android:layout_width="wrap_content"
-            android:transcriptMode="alwaysScroll"
-            android:layout_height="match_parent"
-             />
-
-
-        <RelativeLayout
-            android:id="@+id/llSend"
-            android:layout_alignParentBottom="true"
-            android:layout_width="match_parent"
+        <Button
+            android:id="@+id/btSend"
+            android:layout_width="0dp"
+            android:layout_weight="0.3"
             android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            >
+            android:text="@string/send" />
 
-            <EditText
-                android:id="@+id/etMessage"
-                android:layout_toLeftOf="@+id/btSend"
-                android:layout_alignBottom="@+id/btSend"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/message_hint" />
+    </LinearLayout>
 
-            <Button
-                android:id="@+id/btSend"
-                android:layout_alignParentRight="true"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center_vertical|right"
-                android:text="@string/send" />
-
-        </RelativeLayout>
-
-    </RelativeLayout>
+</RelativeLayout>
